Added "affectAdditionalFieldsUnderPos"
